+++ tests/test_data/test_submodule/f_output_with_ellipsis.m
@@ -1,0 +1,7 @@
+function [output, ...
+    with, ellipsis] = f_output_with_ellipsis(arg)
+% Tests a function with ellipsis in the output
+
+output = 'output';
+with = 'with';
+ellipsis = 'ellipsis';