+++ tests/test_data/test_submodule/super_from_diff_mod.m
@@ -1,0 +1,9 @@
+classdef super_from_diff_mod < MyAbstractClass & MyHandleClass
+    % class which inherits bases from a different module
+    properties
+    end
+    methods
+        function sfdm = super_from_diff_mod()
+        end
+    end
+end